+++ src/main/java/service/PreviewService.java
@@ -21,15 +21,15 @@ public class PreviewService {
     private final static Color DEFAULT = new Color(206, 200, 200, 2);
     private final static Border BORDER = BorderFactory.createLineBorder(DEFAULT, 2);
 
-    public static ArrayList<Long> searchWord(ArrayList<Book> books, int selectedBookIndex, String[] word, DefaultTableModel model){
-        ArrayList<Long> bookIdList = new ArrayList<>();
+	public static ArrayList<Integer> searchWord(ArrayList<Book> books, int selectedBookIndex, String[] word, DefaultTableModel model) {
+		ArrayList<Integer> bookIdList = new ArrayList<>();
 
         int count = 1;
         model.setRowCount(0);
 
         if (selectedBookIndex == 0){ //search in all books
             for (Book book : books){
-                ArrayList<Long> idList;
+	            ArrayList<Integer> idList;
                 for (int i = 0 ; i < word.length ; i++ ) {
                     idList = PreviewService.addLocations
                             (count, model, WordService.findWordInBooks(word[i], book.getId()), book);
@@ -86,9 +86,9 @@ public class PreviewService {
         return preview;
     }
 
-    public static ArrayList<Long> addLocations(int count, DefaultTableModel model, List<WordLocation> locations, Book book) {
+	public static ArrayList<Integer> addLocations(int count, DefaultTableModel model, List<WordLocation> locations, Book book) {
         int index = count;
-        ArrayList<Long> idList = new ArrayList<>();
+		ArrayList<Integer> idList = new ArrayList<>();
 
         for (WordLocation location : locations) {
             model.addRow(new Object[]{index++ , book.getTitle(), book.getAuthor(),
@@ -99,7 +99,7 @@ public class PreviewService {
         return idList;
     }
 
-    public static void createPreview(JTextArea context, String[] word, long bookId, int paragraph){
+	public static void createPreview(JTextArea context, String[] word, int bookId, int paragraph) {
         String text = WordService.buildPreview(bookId,paragraph);
         context.setText(text);
 
@@ -134,7 +134,7 @@ public class PreviewService {
         }
     }
 
-    public static void createPhrasePreview(JTextArea context, String[] phrase, long bookId, int paragraph) {
+	public static void createPhrasePreview(JTextArea context, String[] phrase, int bookId, int paragraph) {
         String text = WordService.buildPreview(bookId, paragraph);
         context.setText(text);
